نمایش اشخاص حقیقی و حقوقی، اعشاری کردن عدد کل سرمایه،

+++ ViewModels/Payment/InvestmentSummaryViewModel.cs
@@ -10,7 +10,7 @@ namespace ViewModels
   public  class InvestmentSummaryViewModel
     {
         [Display(Name = "میزان سرمایه های جذب شده")]
-        public long AmountCapitalRaised { get; set; }
+        public double AmountCapitalRaised { get; set; }
         public int ActiveUsers { get; set; }
         public int InvestmentSuccessPlanCount { get; set; }
         public int? InvestmentCountPerson { get; set; }
